requirements/dev.txt
-r prod.txt
pylint
autopep8
djlint
pylint-django
pylint-celery
flower
requests
